requirements.txt
Flask==3.0.0
google-cloud-firestore==2.13.1
python-dotenv==1.0.0

